+++ wrappers/node/src/api/vcx-base-with-state.ts
@@ -1,7 +1,7 @@
 import * as ffi from 'ffi-napi';
 import { VCXInternalError } from '../errors';
 import { createFFICallbackPromise, ICbRef } from '../utils/ffi-helpers';
-import { Connection } from './connection';
+import { Connection } from './mediated-connection';
 import { VCXBase } from './vcx-base';
 
 export abstract class VCXBaseWithState<SerializedData, StateType> extends VCXBase<SerializedData> {
