+++ wrappers/node/src/index.ts
@@ -2,7 +2,7 @@ export * from './api/init';
 export * from './api/credential-def';
 export * from './api/revocation-registry';
 export * from './api/common';
-export * from './api/connection';
+export * from './api/mediated-connection';
 export * from './api/vcx-mock';
 export * from './api/issuer-credential';
 export * from './api/proof';
